This is turning into a large refactor; major WIP so far to start breaking things apart, fixing the way things are exported and being more consistent across the codebase
