+++ server/src/auth/getPersonTypeFromClaims.ts
@@ -1,7 +1,7 @@
 import { USER_TYPES } from "../constants";
 import { log } from "../log";
 import type { UserType } from "../types";
-import type { AuthorizationClaims } from "./auth.util";
+import type { AuthorizationClaims } from ".";
 
 export function getPersonTypeFromClaims(claims: AuthorizationClaims): UserType {
   const claimsRoles = new Set(claims.role.split(",").map((role) => role.trim()));
