+++ server/src/auth/index.ts
@@ -1,6 +1,8 @@
-export { type GraphQLContext } from "./auth.util";
-export { type ContextUser } from "./userContext";
-export {
-  buildAuthorizationFilter,
-  type PermissionFilters,
-} from "./buildAuthorizationFilter";
+export type { AuthorizationClaims, GraphQLContext } from "./auth.util";
+export type { ContextUser } from "./user";
+export type { PermissionFilters } from "./buildAuthorizationFilter";
+
+export { buildAuthorizationFilter } from "./buildAuthorizationFilter";
+export { decodeToken } from "./decodeToken";
+export { getPersonTypeFromClaims } from "./getPersonTypeFromClaims";
+export { buildContextFromClaims, validateClaims } from "./auth.util";
