+++ server/src/model/person/queries/index.ts
@@ -1,3 +1,4 @@
+export { insertPerson } from "./insertPerson";
 export { selectPerson } from "./selectPerson";
 export { selectPersonOrThrow } from "./selectPersonOrThrow";
 export { selectManyPeople } from "./selectManyPeople";
