+++ server/src/model/rolePermission/queries/index.ts
@@ -1,0 +1,1 @@
+export { selectManyRolePermissions } from "./selectManyRolePermissions";
