+++ server/src/model/state/stateResolvers.test.ts
@@ -2,8 +2,7 @@ import { describe, expect, it, vi } from "vitest";
 import { stateResolvers } from "./stateResolvers";
 import { State as PrismaState } from "@prisma/client";
 import { getManyDemonstrations } from "../demonstration";
-import { ContextUser } from "../../auth/userContext";
-import { GraphQLContext } from "../../auth/auth.util";
+import { ContextUser, GraphQLContext } from "../../auth";
 
 const mockUser = {} as unknown as ContextUser;
 const mockContext: GraphQLContext = {
