+++ server/src/model/systemRoleAssignment/queries/index.ts
@@ -1,0 +1,1 @@
+export { insertSystemRoleAssignment } from "./insertSystemRoleAssignment";
