+++ server/src/model/user/queries/index.ts
@@ -1,3 +1,4 @@
+export { insertUser } from "./insertUser";
 export { selectUser } from "./selectUser";
 export { selectUserOrThrow } from "./selectUserOrThrow";
 export { selectManyUsers } from "./selectManyUsers";
