+++ server/src/model/user/userHistory.prisma
@@ -1,13 +1,15 @@
 model UserHistory {
-  revisionId     Int          @id @default(autoincrement()) @map("revision_id")
-  revisionType   RevisionType @map("revision_type")
-  modifiedAt     DateTime     @default(now()) @map("modified_at") @db.Timestamptz
-  id             String       @db.Uuid
-  personTypeId   String       @map("person_type_id")
-  cognitoSubject String       @map("cognito_subject") @db.Uuid
-  username       String
-  createdAt      DateTime     @map("created_at") @db.Timestamptz
-  updatedAt      DateTime     @map("updated_at") @db.Timestamptz
+  revisionId         Int          @id @default(autoincrement()) @map("revision_id")
+  revisionType       RevisionType @map("revision_type")
+  modifiedAt         DateTime     @default(now()) @map("modified_at") @db.Timestamptz
+  id                 String       @db.Uuid
+  personTypeId       String       @map("person_type_id")
+  cognitoSubject     String?      @map("cognito_subject") @db.Uuid
+  username           String?
+  isMigratedFromPmda Boolean      @map("is_migrated_from_pmda")
+  hasLoggedIn        Boolean      @map("has_logged_in")
+  createdAt          DateTime     @map("created_at") @db.Timestamptz
+  updatedAt          DateTime     @map("updated_at") @db.Timestamptz
 
   @@map("users_history") // user is a reserved word in PostgreSQL; users is not, which is why this is plural
 }
